requirements.txt
fastapi
uvicorn
pillow
python-multipart
pydantic
aiofiles
